Inscriptions: Create pickers for background and punks attributes
The basic UI/UX here is to load the inscriptions (either an iframe or an
image depending on what type this is) into the UI at left and allow the
user to select it.  State is stored at the top-level and injected into
the components.

At the base level each picker takes a list of name/inscription ID pairs
and loads each into the view.

NOTE: Credit to `ordinals-wallet` for providing a full list of the
BitcoinPunks collection.

NOTE: `ordinals.com/content` appears to be a better CDN than
`ord.io/content`, though that may need to change in the future.

[ Testing: npm run build, npm run dev ]