Bitwell Punks Preview: Put up a preview HTML iframe
We generate a preview HTML `<iframe>` tag using the `data:` URL scheme
and base64-encoded HTML string.

The result is a preview of the punk on the background.  Note, we also
use a default background here to avoid any funkiness with the iframe not
loading.

NOTE: The preview base href helps with CORS on localhost previews but
this change does not do any production-level compression for reducing
the inscription size.

[ Testing: npm run build, npm run dev ]