Picker Pagination: Enable a quick pagination button
The DOM was failing to load simply due to its size.  The issue here is
not image optimization but the HTML size alone, which was causing
`view-source` even to not load in localhost.  This change adds a simple
pagination feature so that you can actually load and move through the
different inscription options.

[ Testing: npm run build, npm run dev ]