/api/order: Add ability to mint and POST route for doing so
The basic UI here is just to include the wallet provider options (Hiro,
Unisat, XVerse).  No real major changes (some minor modularity).  The
real algorithm is when the "MINT NOW" button is pressed a POST order is
created and sent to `/api/order` to:

1. Check if the Punk is not reserved
2. Validate payment/recipient addresses
3. Submit a new order request for the base64-encoded HTML to OrdinalsBot
4. Create an active Punk reservation in the database
5. Return the data to the user

Then, the UX just does a simple prompt to pay the sat amount to the
specified OrdinalsBot inscription address.

[ Testing: npm run build, npm run dev ]