Password Protection: Enable password protection (optional) through `crypto-js`
We recently inscribed the min-file for CryptoJS
(https://cdnjs.cloudflare.com/ajax/libs/crypto-js/4.1.1/crypto-js.min.js)
on Bitcoin ordinals, so now we can load it (note it requires a <script>
tag due to the use of the `this` keyword).

When a user puts in an optional password it pops up at the top as an
input field.  Then, once they type it successfully and hit "decrypt" it
disappears.

[ Testing: npm run build, npm run dev ]