fix(/api/cron/update): Respect "paid" field instead of inscribedCount
We have seen the OrdinalsBot API back up even when it detects a user has
paid.  So, instead of checking the `inscribedCount`, honor any user that
pays within the expiration period.

NOTE: The inscription is now likely undefined, so we will need to add a
new route for getting those inscription IDs.

[ Testing: npm run build, npm run dev, deploy ]